Fix. Build Error

+++ Assets/YongJoon/ADBS_Scripts/RE_Gun.cs
@@ -10,7 +10,7 @@ using UnityEngine.Rendering.Universal;
 public abstract class RE_Gun : MonoBehaviour
 {
     [SerializeField] protected Transform muzzlePos;
-    [SerializeField] protected float fireDamage;
+    [SerializeField] protected int fireDamage;
     [SerializeField] protected int allBullet;             // 총탄 개수
     [SerializeField] protected int availableBullet;       // 재장전 하기 전 사용 가능한 총탄 개수 
     [SerializeField] protected int curAvailavleBullet;    // 계속 사용 중인 현재 총알 개수
